backend/requirements.txt
# AI Oral Exam - Backend Dependencies
# ====================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Firebase Admin SDK
firebase-admin>=6.4.0

# LLM APIs (unified interface)
litellm>=1.30.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# Voice/Audio & Real-time Communication
websockets>=12.0
google-cloud-speech>=2.24.0
google-cloud-texttospeech>=2.16.0
# pyaudio>=0.2.14  # Commented out - requires system dependencies
# Additional Gemini Live support
# google.generativeai>=0.8.0  # Conflicts with google-generativeai

# Data validation
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0

# Async support
httpx>=0.26.0
aiofiles>=23.2.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0  # Retry logic for API calls
numpy>=1.24.0  # For vector operations in RAG
